[FEAT/#344] 추천 친구뷰 / 프로필 추가

+++ data/src/main/java/com/example/data/repository/ProfileRepositoryImpl.kt
@@ -28,7 +28,7 @@ class ProfileRepositoryImpl @Inject constructor(
         }
     }
 
-    override suspend fun deleteFriendData(friendId: Int): Result<Unit> {
+    override suspend fun deleteFriendData(friendId: Long): Result<Unit> {
         return runCatching {
             profileDataSource.deleteFriendData(friendId).data
         }
